+++ crates/backend-uzu/src/config/weaver.rs
@@ -4,14 +4,14 @@ use crate::config::{linear::LinearConfig, normalization::NormalizationConfig, ro
 
 #[uzu_config]
 pub struct WeaverConfig {
-    pub model_dim: usize,
-    pub target_model_dim: usize,
-    pub target_embedding_dim: usize,
-    pub num_layers: usize,
-    pub num_heads: usize,
-    pub hidden_dim: usize,
-    pub max_depth: usize,
-    pub candidate_pool_size: usize,
+    pub model_dim: u32,
+    pub target_model_dim: u32,
+    pub target_embedding_dim: u32,
+    pub num_layers: u32,
+    pub num_heads: u32,
+    pub hidden_dim: u32,
+    pub max_depth: u32,
+    pub candidate_pool_size: u32,
     pub linear_config: LinearConfig,
     pub norm_config: NormalizationConfig,
     pub rope_config: AnyRoPEConfig,
